+++ packages/lambda/scripts/latest-versions.sh
@@ -6,11 +6,28 @@
 #
 # Requires jq
 #
-# Usage: ./latest-versions.sh
+# Usage: ./latest-versions.sh [chromium|firefox]
 #
 
-CHROMIUM=$(curl -s https://omahaproxy.appspot.com/all.json | \
-  jq -r '.[] | select(.os == "linux") | .versions[] | select(.channel == "stable") | .current_version' \
-)
+set -e
 
-echo "Current stable version of Chromium: $CHROMIUM"
+cd "$(dirname "$0")/.."
+
+PACKAGE_DIRECTORY=$(pwd)
+
+version() {
+  BUILD_NAME=$1
+  cd "$PACKAGE_DIRECTORY/builds/$BUILD_NAME"
+  echo "--- $BUILD_NAME ---"
+  ./latest.sh
+}
+
+if [ ! -z "$1" ]; then
+  version "$1"
+else
+  cd "$PACKAGE_DIRECTORY/builds"
+
+  for DOCKER_FILE in */latest.sh; do
+    version "${DOCKER_FILE%%/*}"
+  done
+fi