+++ packages/lambda/scripts/postinstall.js
@@ -6,23 +6,24 @@ const path = require('path')
 const https = require('https')
 const extract = require('extract-zip')
 
-const RELEASE_DOWNLOAD_URL_BASE = 'https://github.com/adieuadieu/serverless-chrome/releases/download'
+const RELEASE_DOWNLOAD_URL_BASE =
+  'https://github.com/adieuadieu/serverless-chrome/releases/download'
 
 function unlink (path) {
   return new Promise((resolve, reject) => {
-    fs.unlink(path, error => error ? reject(error) : resolve())
+    fs.unlink(path, error => (error ? reject(error) : resolve()))
   })
 }
 
 function rename (from, to) {
   return new Promise((resolve, reject) => {
-    fs.rename(from, to, error => error ? reject(error) : resolve())
+    fs.rename(from, to, error => (error ? reject(error) : resolve()))
   })
 }
 
 function extractFile (file, destination) {
   return new Promise((resolve, reject) => {
-    extract(file, { dir: destination }, error => error ? reject(error) : resolve())
+    extract(file, { dir: destination }, error => (error ? reject(error) : resolve()))
   })
 }
 
@@ -40,7 +41,7 @@ function get (url, destination) {
         }
 
         const file = fs.createWriteStream(destination)
-  
+
         response.pipe(file)
 
         file.on('finish', () => {
@@ -55,26 +56,26 @@ function get (url, destination) {
 }
 
 function getChromium () {
-  const ZIP_FILENAME = `headless-chromium-${process.env.npm_package_config_chromiumVersion}-amazonlinux-2017-03.zip`
+  const ZIP_FILENAME = `headless-chromium-${process.env
+    .npm_package_config_chromiumVersion}-amazonlinux-2017-03.zip`
   const ZIP_URL = `${RELEASE_DOWNLOAD_URL_BASE}/${process.env.npm_package_version}/${ZIP_FILENAME}`
   const DOWNLOAD_PATH = path.resolve(__dirname, '..', ZIP_FILENAME)
   const EXTRACT_PATH = path.resolve(__dirname, '..', 'dist')
 
   if (fs.existsSync(DOWNLOAD_PATH)) {
-    console.log(`Precompiled headless Chromium binary for AWS Lambda previously downloaded. Skipping download.`)
+    console.log('Precompiled headless Chromium binary for AWS Lambda previously downloaded. Skipping download.')
     return Promise.resolve()
   }
 
   console.log('Downloading precompiled headless Chromium binary for AWS Lambda.')
-  
+
   return get(ZIP_URL, DOWNLOAD_PATH)
     .then(() => extractFile(DOWNLOAD_PATH, EXTRACT_PATH))
     .then(() => console.log('Completed Headless Chromium download.'))
 }
 
 if (require.main === module) {
-  getChromium()
-  .catch(error => {
+  getChromium().catch((error) => {
     console.error(error)
     process.exit(1)
   })