+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - @serverless-chrome/lambda NPM package
 - serverless-plugin-chrome NPM package for Serverless-framework
 - CHANGELOG.md.
+- Dockerfile
+- build and release tooling
+- Headless Chrome headless_shell binary version 64.0.3242.2 built for AWS Lambda 
 
 ### Changed
 - example Serverless-framework printToPdf function handler to use the Serverless plugin